requirements.txt
affine==2.3.0
amqp==5.0.3
asgiref==3.3.0
attrs==20.3.0
autopep8==1.5.4
billiard==3.6.3.0
branca==0.4.1
celery==5.0.5
certifi==2020.6.20
cftime==1.3.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.0
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.0.1
cycler==0.10.0
deprecation==2.1.0
dj-database-url==0.5.0
Django==3.1.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.1
drf-yasg==1.20.0
Fiona==1.8.17
folium==0.11.0
GDAL==3.0.4
geopandas==0.8.1
gunicorn==20.0.4
h5netcdf==0.8.1
h5py==3.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.3.1
kombu==5.0.2
MarkupSafe==1.1.1
matplotlib==3.3.3
munch==2.5.0
netCDF4==1.5.4
numpy==1.19.4
packaging==20.8
pandas==1.1.4
Pillow==8.0.1
prompt-toolkit==3.0.14
pycodestyle==2.6.0
pycosat==0.6.3
pygeos==0.8
pyparsing==2.4.7
pyproj==3.0.0.post1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
rasterio==1.1.8
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
SQLAlchemy==1.3.20
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
wradlib==1.9.0
xarray==0.16.2
xlrd==1.2.0
xmltodict==0.12.0
